pyproject.toml
[tool.poetry]
name = "multimodal_rag"
version = "0.1.0"
description = "Multimodal Retrieval-Augmented Generation pipeline"
authors = ["Alex Zhukov"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/zhukov-alex/multimodal-rag"
packages = [{ include = "multimodal_rag", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "3.9.5"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anyio = "4.9.0"
attrs = "25.3.0"
beautifulsoup4 = "4.13.4"
certifi = "2025.4.26"
chardet = "4.0.0"
charset-normalizer = "3.4.1"
click = "8.1.8"
frozenlist = "1.6.0"
h11 = "0.16.0"
httpcore = "1.0.9"
httpx = "0.28.1"
idna = "3.10"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
langchain-core = "0.3.57"
langchain-text-splitters = "0.3.8"
langdetect = "1.0.9"
langsmith = "0.3.40"
lxml = "5.4.0"
markdown-it-py = "3.0.0"
markdownify = "0.13.1"
mdurl = "0.1.2"
multidict = "6.4.3"
orjson = "3.10.18"
packaging = "24.2"
propcache = "0.3.1"
pydantic = "2.11.4"
pydantic-core = "2.33.2"
pygments = "2.19.1"
pypdf = "5.4.0"
python-docx = "1.1.2"
python-dotenv = "1.1.0"
python-json-logger = "2.0.7"
pyyaml = "6.0.2"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rich = "14.0.0"
shellingham = "1.5.4"
six = "1.17.0"
sniffio = "1.3.1"
soupsieve = "2.7"
tenacity = "9.1.2"
tqdm = "4.67.1"
typer = "0.15.3"
typing-inspection = "0.4.0"
typing-extensions = "4.13.2"
urllib3 = "2.4.0"
yarl = "1.20.0"
zstandard = "0.23.0"
pillow = "^11.2.1"
weaviate-client = "^4.7.0"

[tool.poetry.dev-dependencies]
black = "^24.0"
coverage = "^7.0"
flake8 = "^6.0"
isort = "^5.0"
mypy = "^1.0"
pre-commit = "^3.0"
pytest = "^7.0"

[tool.poetry.scripts]
index-pipeline = "multimodal_rag.cli:app"
rag-pipeline = "multimodal_rag.cli:app"

[tool.mypy]
python_version = "3.11"
strict = true
exclude = ["tests"]
